+++ src/domain/RacingCarGame.js
@@ -1,4 +1,4 @@
-const { GAME_RULE } = require('../constants/rule');
+const { GAME_RULE } = require('./constants/rule');
 const pickNumberInRange = require('../utils/pickNumberInRange');
 
 class RacingCarGame {
