+++ src/domain/constants/rule.js
